refactor: OAuth 관련 파일 및 경로 정리
