--- apps/client/src/features/auth/api/end-point.ts
@@ -1,5 +1,0 @@
-export const END_POINT = {
-  LOGIN: {
-    EXCHANGE_GOOGLE_CODE: 'oauth/google/callback',
-  },
-};
