+++ apps/client/src/pages/login/apis/end-point.ts
@@ -1,3 +1,4 @@
 export const LOGIN_END_POINT = {
   LOGIN: 'oauth/google',
+  EXCHANGE_GOOGLE_CODE: 'oauth/google/callback',
 };
