+++ apps/client/src/pages/login/components/login-callback/index.ts
